backend/pyproject.toml
[tool.poetry]
name = "english_dictionary"
version = "0.1.0"
description = ""
authors = ["Andrew Klyuchnikov <off-rey@bk.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
